tooling.py: rename as travistooling.py for disambiguation
