+++ .travis/run_task.py
@@ -27,7 +27,7 @@ import sys
 
 from should_publish import should_publish
 from should_run_tests import should_run_tests
-from tooling import make, rreplace
+from travistooling import make, rreplace
 
 
 def main():
